using graphql-import and cleanup codes
