+++ node/src/typedefs.js
@@ -1,0 +1,6 @@
+const path = require('path')
+const { importSchema } = require('graphql-import')
+
+const typeDefs = importSchema(path.resolve(__dirname + '/schema.graphql'))
+
+module.exports = { typeDefs }
